algobio-ai-classifier/requirements.txt
fastapi
uvicorn
joblib
numpy
scikit-learn
datasets
python-multipart

sqlalchemy
psycopg2-binary
pydantic-settings
matplotlib
pandas